+++ ocr-correction-viklofg-sweocr/tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from sbx_ocr_correction_viklofg_sweocr import (
+from sbx_ocr_correction_viklofg_sweocr.ocr_corrector import (
     OcrCorrector,
 )
